Refactor Proxy into GovukProxy and GovukProxyProfiles and runner
This commit renames `proxy.rb` to `govuk_proxy.rb` (and renames
from the old `MyProxy` to `GovukProxy`). It then splits the proxy
into a `govuk_proxy_runner.rb` (which starts the proxy on port
8080) and `features/support/govuk_proxy.rb` (which has the logic
for the proxy itself). The proxy profiles have been split out
into a `govuk_proxy_profiles.rb` module in the same `support` dir.

Benefits:

- Single responsibility principle: adding a new proxy profile
  means editing one self-contained file. The runner runs the proxy.
  The proxy logic lives in govuk_proxy.
- Better validation: we now raise an exception in both the proxy
  runner and in `env.rb` if a bad `GOVUK_PROXY_PROFILE` is provided.
  This validation used to only happen in the proxy runner. This
  should catch potential misconfiguration issues.
- The proxy profiles and proxy logic live in the `support` folder
  without making the shell hang on starting up a server on port 8080
  (which now is the sole responsibility of the proxy runner).
  This makes it theoretically possible to reuse parts of the proxy
  in our step definitions, perhaps making it possible to kill off
  the `http_requests.rb` file that predates the proxy.
