+++ README.md
@@ -55,7 +55,7 @@ You can use the following environment variables to configure the tests:
 * `SIGNON_EMAIL`: email of a Signon user in $ENVIRONMENT
 * `SIGNON_PASSWORD`: password of a Signon user in $ENVIRONMENT
 * `RATE_LIMIT_TOKEN`: (optional) a token used to bypass rate limiting if present on apps.
-* `PROXY_PROFILE` (optional) name of profile to use in proxy.rb. See [proxy tests](#proxy)
+* `GOVUK_PROXY_PROFILE` (optional) name of profile to use in govuk_proxy.rb. See [proxy tests](#proxy)
 
 Smokey cannot use a Signon account which has multi-factor authentication
 enabled.
@@ -70,28 +70,28 @@ k get secret smokey-signon-account -oyaml | yq .data.email | base64 -d
 
 ### Proxy tests
 
-The Smokey test suite can be told to run against our primary CDN (Fastly), our failover CDN (Cloudfront), or any one of the GOV.UK mirrors. To do so, you'll need to run the proxy in a separate process, and ensure you specify the `PROXY_PROFILE` environment variable.
+The Smokey test suite can be told to run against our primary CDN (Fastly), our failover CDN (Cloudfront), or any one of the GOV.UK mirrors. To do so, you'll need to run the proxy in a separate process, and ensure you specify the `GOVUK_PROXY_PROFILE` environment variable.
 
 Example of running Smokey tests against the main GOV.UK mirror:
 
 ```shell
-shell1$ env PROXY_PROFILE=mirrorS3 ruby proxy.rb
+shell1$ env GOVUK_PROXY_PROFILE=mirrorS3 ruby govuk_proxy_runner.rb
 
-shell2$ env PROXY_PROFILE=mirrorS3 ENVIRONMENT=production bundle exec cucumber --tags="@worksonmirror"
+shell2$ env GOVUK_PROXY_PROFILE=mirrorS3 ENVIRONMENT=production bundle exec cucumber --tags="@worksonmirror"
 ```
 
 Example of running Smokey tests against the failover CDN (NB: the 'secret' Cloudfront value can be [found in govuk-dns-tf](https://github.com/alphagov/govuk-dns-tf/pull/69)):
 
 ```shell
-shell1$ env PROXY_PROFILE=failoverCDN FAILOVER_CDN_HOST="secret.cloudfront.net" ruby proxy.rb
+shell1$ env GOVUK_PROXY_PROFILE=failoverCDN FAILOVER_CDN_HOST="secret.cloudfront.net" ruby govuk_proxy_runner.rb
 
-shell2$ env PROXY_PROFILE=failoverCDN ENVIRONMENT=production bundle exec cucumber --tags="not @notreplatforming and not @notcloudfront"
+shell2$ env GOVUK_PROXY_PROFILE=failoverCDN ENVIRONMENT=production bundle exec cucumber --tags="not @notreplatforming and not @notcloudfront"
 ```
 
 Debug mode for checking you're setting headers correctly:
 
 ```shell
-env PROXY_PROFILE=debug ruby proxy.rb
+env GOVUK_PROXY_PROFILE=debug ruby govuk_proxy_runner.rb
 # Visit http://127.0.0.1:8080/ to see your request headers
 ```
